Preserve addon configuration when deploying
The configuration is stored in `meta.json` file, which exists
inside the addon folder. When deploying the whole folder is deleted,
and then recreating, which also removes the configuration.

Instead, move the configuration file to addons folder,
before deleting the plugin's folder, and move it back to the
plugin's folder after it is recreated.